JH-165 Refactor: Add NoteData,TopicData,PageData class for remove duplicate
